backend/requirements.txt
astroid
Click
ecdsa
Flask
Flask-SQLAlchemy
future==0.17.1
isort==4.3.18
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six
typed-ast==1.4.2
Werkzeug
wrapt==1.11.1
Flask-Cors
